Irrigation - goGetValveTimes() spamming controller
The goGetValveTimes functions was sending 3 requests to the controller without pause or waiting for response - thus the receive code was rejecting the replies and the sender was NACKing some sends. This update replaces the goGetValveTimes function with a timed receiver that works 100% of the time. Other changes to make the VALVE_TIMES_RELOAD a variable with comments at the top and adjustments to the Setup to call this updated method.