Bug with manual button not sending status
This mod fixes an issue where the status of the system was NOT sent to the controller on a manual button operation due to the lastState being updated to a new state in the displayMenu. The displayMenu was called just before the lastState = state; line which meant the state was never iterated with the changed value from the displayMenu. The relays would click but the status was never transmitted.

Also added some minor 50ms delays on the status updates to reduce radio saturation. Changed some DEBUG messages to show the Zone names on time change and made the VAR_3 display consistent with other VAR debug displays.