New version
I'll leave it to you if this code should replace the old code. It has more features, so its bigger.
It's the same code I just added to the weather station example.

The idea with both is that it will help people avoid old sensors and lots of work.

Long live the BME280.